requirements.txt
Django==2.2.2
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
requests==2.22.0
rq==1.0
